move compute_sample_durations
