+++ doc/api.rst
@@ -178,7 +178,6 @@ Utilities related to spatial data.
    :toctree: generated/
 
    compute_nbins
-   compute_sample_durations
    compute_bin_width
    compute_pos_ranges
    convert_2dindices
@@ -507,6 +506,7 @@ Utilities for working with timestamps.
 .. autosummary::
    :toctree: generated/
 
+   compute_sample_durations
    infer_time_unit
    convert_ms_to_sec
    convert_sec_to_min
