+++ spiketools/tests/utils/test_timestamps.py
@@ -7,6 +7,16 @@ from spiketools.utils.timestamps import *
 ###################################################################################################
 ###################################################################################################
 
+def test_compute_sample_durations():
+
+    # define a timestamp, with irregular times
+    timestamp = np.array([0.0, 1.0, 2.0, 3.0, 4.5, 5.0, 6.0, 7.0, 8.0, 9.0, 12.0])
+    bin_time = compute_sample_durations(timestamp)
+
+    # check dimensions & sum
+    assert bin_time.shape[0] == timestamp.shape[0]
+    assert np.sum(np.diff(timestamp)) == np.sum(bin_time)
+
 def test_infer_time_unit(tspikes):
 
     # Check test data in seconds
